Start moving prometheus module hist handling to libnoit
